add svg for resume download, remove unused footer elements, css cleanup, svg dir renamed to images
