+++ src/App.css
@@ -164,14 +164,14 @@ a:hover .gallery-text {
     transition-timing-function: ease-in-out;
 }
 
-img {
+.main-container img {
     opacity: 0.8;
     transition: opacity 0.5s;
     transition-delay: 0.25s;
     transition-timing-function: ease-in-out;
 }
 
-a:hover img {
+.main-container a:hover img {
     opacity: 0;
     transition: opacity 0.25s;
     transition-timing-function: ease-in-out;
