+++ src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import pfp from './svg/pfp.png';
+import pfp from './images/pfp.png';
 import Projects from './components/Projects';
 import Socials from './components/Socials';
 import ParticlesBackground from './components/ParticlesBackground';
@@ -10,10 +10,10 @@ import Footer from './components/Footer';
 /*
 TODO:
 
-Can I make this like fullpage.js?
-Sect 1 = Title page
-Sect 2 = Projects
-Sect 3 = Contributions
+Change element IDs to classes
+Fill out footer with content
+Upload resume to S3
+Add email to footer
 */
 function App() {
   const galleryHead = useRef(null);
@@ -45,17 +45,12 @@ function App() {
       <header className="App-header">
         <div id="header-text">
             <h1 id="main-header">NIKO RAISANEN</h1>
-            {/* <h5>Full-Stack Developer and AppSec Engineer</h5> */}
         </div>
 
         <img src={pfp} id="header-pic" alt="headshot"/>
         <Socials/>
         
         <div id="intro">
-            {/* <p className="headline">Full-Stack Developer and AppSec Engineer</p> */}
-            {/* <p>I have many titles - Software Engineer, AppSec Engineer, Fisherman, but first and foremost, I am a creative problem solver.</p> */}
-            {/* <p>Hi there, I'm Niko! I'm a software engineer with a foundation in security engineering, always eager to dive in and bring my ideas to life through code. I enjoy brainstorming architectural decisions to come up with the most efficient solution on both the infrastructure and code levels. When I'm not in front of a keyboard you can find me fishing or snowboarding</p> */}
-            {/* <p>Hi there, I'm Niko! I'm a software engineer with a background in security engineering, always eager to dive in and bring my ideas to life through code</p> */}
             <p>Hello! I'm Niko, a software engineer with a background in security engineering. I have a passion for turning my ideas into tangible solutions through code and am always eager to dive into a new challenge.</p>
         </div>
       </header>
