+++ src/footer.css
@@ -1,125 +1,95 @@
-  footer {
-    margin-top: 30%;
-    background: #1a1a21fc;
-    box-shadow: 0px 10px 40px 0px #242126;
-    /* 242126 */
-  
-    height: auto;
-  
-    width: 100vw;
-  
-    padding-top: 40px;
-  
-    color: #fff;
+footer {
+  margin-top: 30%;
+  background: #1a1a21fc;
+  box-shadow: 0px 10px 40px 0px #242126;
+  height: auto;
+  width: 100vw;
+  padding-top: 40px;
+  color: #fff;
+}
+
+.footer-content {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  text-align: center;
+}
+
+.footer-content h3 {
+  font-size: 2.1rem;
+  font-weight: 500;
+  text-transform: capitalize;
+  line-height: 3rem;
+  margin-bottom: 25px;
+}
+
+.footer-content p {
+  max-width: 500px;
+  margin: 10px auto;
+  line-height: 28px;
+  font-size: 14px;
+  color: #cacdd2;
+}
+
+.socials {
+  list-style: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin: 1rem 0 3rem 0;
+}
+
+.socials li {
+  margin: 0 10px;
+}
+
+.socials a {
+  text-decoration: none;
+  color: #fff;
+  padding: 5px;
+}
+
+.socials a i {
+  font-size: 1.1rem;
+  width: 20px;
+  transition: color 0.4s ease;
+}
+
+.socials a:hover i {
+  color: aqua;
+}
+
+.footer-menu {
+  float: right;
+}
+
+.footer-menu ul {
+  display: flex;
+}
+
+.footer-menu ul li {
+  padding-right: 10px;
+  display: block;
+}
+
+.footer-menu ul li a {
+  color: #cfd2d6;
+  text-decoration: none;
+}
+
+.footer-menu ul li a:hover {
+  color: #27bcda;
+}
+
+@media (max-width: 500px) {
+  .footer-bottom p {
+    float: none;
   }
-  
-  .footer-content {
-    display: flex;
-  
-    align-items: center;
-  
-    justify-content: center;
-  
-    flex-direction: column;
-  
-    text-align: center;
-  }
-  
-  .footer-content h3 {
-    font-size: 2.1rem;
-  
-    font-weight: 500;
-  
-    text-transform: capitalize;
-  
-    line-height: 3rem;
-  }
-  
-  .footer-content p {
-    max-width: 500px;
-  
-    margin: 10px auto;
-  
-    line-height: 28px;
-  
-    font-size: 14px;
-  
-    color: #cacdd2;
-  }
-  
-  .socials {
-    list-style: none;
-  
-    display: flex;
-  
-    align-items: center;
-  
-    justify-content: center;
-  
-    margin: 1rem 0 3rem 0;
-  }
-  
-  .socials li {
-    margin: 0 10px;
-  }
-  
-  .socials a {
-    text-decoration: none;
-  
-    color: #fff;
-  
-    padding: 5px;
-  }
-  
-  .socials a i {
-    font-size: 1.1rem;
-  
-    width: 20px;
-  
-    transition: color 0.4s ease;
-  }
-  
-  .socials a:hover i {
-    color: aqua;
-  }
-  
-  .footer-menu {
-    float: right;
-  }
-  
   .footer-menu ul {
     display: flex;
+    margin-top: 10px;
+    margin-bottom: 20px;
+    text-align: center;
   }
-  
-  .footer-menu ul li {
-    padding-right: 10px;
-  
-    display: block;
-  }
-  
-  .footer-menu ul li a {
-    color: #cfd2d6;
-  
-    text-decoration: none;
-  }
-  
-  .footer-menu ul li a:hover {
-    color: #27bcda;
-  }
-  
-  @media (max-width: 500px) {
-    .footer-bottom p {
-      float: none;
-    }
-  
-    .footer-menu ul {
-      display: flex;
-  
-      margin-top: 10px;
-  
-      margin-bottom: 20px;
-  
-      text-align: center;
-    }
-  }
-  
+}
